Experimental simple-BVH implementation
This is a test to compare a very simple BVH method against
the BentleyOttmann algorithm.

The main down-side is it has a much worse, worst-case,
when all segments are long and overlap roughly the same region.

Although this could be resolved using spatial splits,
float precision would need to be handled carefully in that case.